requirements.txt
pysrt
moviepy
syncedlyrics
pylrc
stable-ts
python_speech_features
levenshtein
mutagen
levenshtein
manim
Pillow==9.4.0
faster-whisper
ffmpeg-python
demucs
fastapi
uvicorn
sse-starlette
